requirements.txt
feedparser
requests
pyyaml
python-dotenv
openai